feat: SMTPAgent to inherit from ServerAgent
